+++ src/main/java/racingcar/domain/Winner.java
@@ -1,4 +1,4 @@
-package racingcar.model;
+package racingcar.domain;
 
 import java.util.List;
 import java.util.stream.Collectors;
