+++ src/main/java/racingcar/service/RandomNumberGenerator.java
@@ -1,4 +1,4 @@
-package racingcar.module;
+package racingcar.service;
 
 import racingcar.Constant;
